+++ datumaro/plugins/data_formats/yolo/format.py
@@ -15,6 +15,11 @@ class YoloLoosePath:
     NAMES_FILE = "obj.names"
 
 
+class YoloUltralyticsPath:
+    META_FILE = "data.yaml"
+
+
 class YoloFormatType(IntEnum):
     yolo_strict = 0
     yolo_loose = 1
+    yolo_ultralytics = 2