+++ tests/unit/data_formats/datumaro/test_datumaro_binary_format.py
@@ -16,6 +16,12 @@ from datumaro.plugins.data_formats.datumaro_binary import (
 )
 from datumaro.plugins.data_formats.datumaro_binary.crypter import Crypter
 from datumaro.plugins.data_formats.datumaro_binary.format import DatumaroBinaryPath
+
+# pylint: disable=undefined-variable
+# We have to disable this because IntListMapper and FloatListMapper are imported by the following line,
+# but Pylint raises the following errors.
+# tests/unit/data_formats/datumaro/test_datumaro_binary_format.py:107:16: E0602: Undefined variable 'IntListMapper' (undefined-variable)
+# tests/unit/data_formats/datumaro/test_datumaro_binary_format.py:115:16: E0602: Undefined variable 'FloatListMapper' (undefined-variable)
 from datumaro.plugins.data_formats.datumaro_binary.mapper import *
 from datumaro.plugins.data_formats.datumaro_binary.mapper.annotation import AnnotationMapper